Finally progress!!
All along just needed to wrap all the %s placeholders in ' ' for the actual VALUES being inserted. Lots of refactoring to update this and will be testing it all tomorrow. Mock table and insert worked as expected with this so look to be moving forward and coding some search query implementations as well.