Changed up and reorganized foreign key constraints for all tables.
Added new columns indicative for each table to have its own gps_shot and using that table gps_shot as the FK to reference the gps_shot in the attributes table. Still need to change the column name in the table inserts but will be setting the value with the same CommonAttributes class variable as the attributes table so that the gps_shots will match, be the same and enable joining.