More Joins.
Inner Join on the bend table is working flawlessly. Need some work on the combo bend however. The query result is only returning the station number at this time and no other data. In postgres, I successfully coded a multiple inner join for the combo bend and mirrored that query in the app in psycopg2 (or thought I did). Will investigate this query.